Move Go code into pkg
