+++ libs/sdk/go/project.json
@@ -2,7 +2,7 @@
     "name": "lib-go",
     "$schema": "../../../node_modules/nx/schemas/project-schema.json",
     "projectType": "library",
-    "sourceRoot": "libs/sdk/go",
+    "sourceRoot": "libs/sdk/go/pkg/buildkite",
     "tags": [],
     "targets": {
         "test": {
